requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bootstrap4==0.1.0
Django==3.2.3
django-bootstrap4==3.0.1
django-filter==2.4.0
Pillow==8.2.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
